requirements.txt
mysql-connector-python==8.0.17
nltk==3.2.5
numpy==1.14.2
pandas==0.22.0
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
tensorflow==1.12.0
tensorflow-plot==0.3.0
tensorflow-probability==0.7.0
